인벤토리 진행중. 아이템 관련 수정중
인벤토리 구현중.
아이템 데이터 관련 수정중

+++ Assets/Scripts/Money.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class Money : MonoBehaviour
 {
-    public float moneyAmount { private set; get; }
+    public int moneyAmount { private set; get; }
 
-    public void SetMoneyAmount(float money)
+    public void SetMoneyAmount(int money)
     {
         moneyAmount = money;
         Debug.Log("Drop Money: "+moneyAmount);
